bug fixed

+++ app/controllers/ManagerController.php
@@ -36,7 +36,9 @@ class ManagerController extends \BaseController {
 		})->get();
 
 		foreach ($havingEqGroups as $g) {
-			$g->users()->detach($userId[0]);
+			if (substr($g->key, 5)==="equip") {
+				$g->users()->detach($userId[0]);
+			}
 		};
 
 		// 선택한 노드의 관리자로 임명함
